requirements.txt
backports.csv==1.0.7
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.13.2
cssselect2==0.2.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.1.9
django-chartjs==1.5.0
django-crispy-forms==1.7.2
django-filter==2.2.0
django-import-export==1.2.0
et-xmlfile==1.0.1
html5lib==1.0.1
jdcal==1.4.1
odfpy==1.4.0
openpyxl==2.6.3
Pillow==6.1.0
pycparser==2.19
Pyphen==0.9.5
pytz==2019.2
PyYAML==5.1.2
six==1.12.0
tablib==0.13.0
tinycss2==1.0.2
WeasyPrint==50
webencodings==0.5.1
xlrd==1.2.0
xlwt==1.3.0
